chords editro cleanup and refactoring
