--- src/components/Editor/EditorPage.scss
@@ -1,7 +1,0 @@
-@import '../../stylesheets/common.scss';
-.editor-page {
-    &__content {
-        padding: $spacing;
-        height: 100%;
-    }
-}
