--- src/components/Editor/TextEditor.scss
@@ -1,6 +1,0 @@
-@import '@stylesheets/common.scss';
-.text-editor {
-    &__content {
-        height: 100%;
-    }
-}
