+++ src/components/Editor/chordsEditor/InstrumentalPart/InstrumentalPart.scss
