+++ src/components/Editor/chordsEditor/InstrumentalPart/InstrumentalPart.tsx
@@ -1,7 +1,8 @@
 import { InstrumentalPartType } from '@model/song';
 import React from 'react';
-import { Chord } from './Chord';
+import { Chord } from '../components/Chord';
 import './InstrumentalPart.scss';
+import { noop } from 'lodash';
 
 const CLASS = 'instrumental-part';
 
@@ -13,7 +14,7 @@ export const InstrumentalPart: React.FC<InstrumentalPartProps> = ({ part }) => {
     return (
         <div className={CLASS}>
             <div>{part.title}</div>
-            <div>
+            <>
                 {part.chords
                     ? part.chords
                           .map((line, index) =>
@@ -22,12 +23,14 @@ export const InstrumentalPart: React.FC<InstrumentalPartProps> = ({ part }) => {
                                   <Chord
                                       key={`${index}_${chordIndex}`}
                                       chord={chord}
+                                      onEdit={noop}
+                                      onRemove={noop}
                                   />
                               ))
                           )
                           .reduce((prev, next) => [...prev, <br />, ...next])
                     : part.tabs}
-            </div>
+            </>
         </div>
     );
 };
