+++ src/components/Editor/chordsEditor/InstrumentalPart/index.ts
@@ -1,0 +1,1 @@
+export * from './InstrumentalPart';
