+++ src/components/Editor/chordsEditor/LyricsPart/Line/EditLine.scss
