+++ src/components/Editor/chordsEditor/LyricsPart/Line/LyricsLine/Letter.scss
