+++ src/components/Editor/chordsEditor/LyricsPart/Line/LyricsLine/LyricsLine.scss
