+++ src/components/Editor/chordsEditor/LyricsPart/Line/LyricsLine/index.ts
