+++ src/components/Editor/chordsEditor/LyricsPart/Line/index.ts
