+++ src/components/Editor/chordsEditor/LyricsPart/LyricsPart.scss
