+++ src/components/Editor/chordsEditor/LyricsPart/LyricsPart.tsx
