+++ src/components/Editor/chordsEditor/LyricsPart/index.ts
@@ -1,0 +1,1 @@
+export * from './LyricsPart';
