+++ src/components/Editor/chordsEditor/index.ts
@@ -1,0 +1,1 @@
+export * from './ChordsEditor';
